requirements.txt
streamlit
python-dotenv
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-chroma
langchain-text-splitters
pypdf